requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.1.0
Django==6.0.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
freetype-py==2.5.1
gunicorn==24.1.1
html5lib==1.1
idna==3.11
lxml==6.0.2
numpy==2.4.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.6.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.9
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
